chore(connect): restructure fw update files
