+++ packages/connect/src/firmware/__tests__/calculateFirmwareHash.test.ts
