+++ packages/connect/src/firmware/__tests__/parseFirmwareHeaders.test.ts
