+++ packages/connect/src/firmware/__tests__/verifyEntropy.test.ts
