+++ packages/connect/src/firmware/calculateFirmwareHash.ts
