+++ packages/connect/src/firmware/getBinary.ts
@@ -2,9 +2,9 @@ import { VersionArray } from '@trezor/device-utils';
 import { removeTrailingSlashes } from '@trezor/utils';
 
 import { parseFirmwareHeaders } from './parseFirmwareHeaders';
-import { ERRORS } from '../../constants';
-import { BinaryInfo } from '../../types';
-import { httpRequest } from '../../utils/assets';
+import { ERRORS } from '../constants';
+import { BinaryInfo } from '../types';
+import { httpRequest } from '../utils/assets';
 
 const MIN_FIRMWARE_SIZE_BYTES = 200;
