+++ packages/connect/src/firmware/index.ts
@@ -1,5 +1,3 @@
 export { getBinary, getBinaryOptional } from './getBinary';
 export { shouldStripFwHeaders, stripFwHeaders } from './modifyFirmware';
-export { uploadFirmware } from './uploadFirmware';
 export { calculateFirmwareHash } from './calculateFirmwareHash';
-export { parseFirmwareHeaders } from './parseFirmwareHeaders';
