+++ packages/connect/src/firmware/parseFirmwareHeaders.ts
