+++ packages/connect/src/firmware/x509certificate.ts
